# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company Ship API",
        "description": "API quản lý vận tải - công ty, nhân sự, xe, chuyến đi, lương",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/offices": {
            "get": {
                "tags": [
                    "Offices"
                ],
                "summary": "Danh sách văn phòng",
                "operationId": "0e67bc605064852df11a2a6f516d7c86",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo code, name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Lọc theo công ty",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sắp xếp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "Offices"
                ],
                "summary": "Tạo văn phòng mới",
                "operationId": "a1b82ec7b5400e06e201998d6ada2f60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "company_id"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "OFF001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Văn phòng HN"
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "manager_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            }
        },
        "/api/offices/{id}": {
            "get": {
                "tags": [
                    "Offices"
                ],
                "summary": "Chi tiết văn phòng",
                "operationId": "f40fa832786feb30bd5645b8acae8480",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "Offices"
                ],
                "summary": "Cập nhật văn phòng",
                "operationId": "eb7b7e79ef22fe0a944e952dbceae4bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "manager_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Offices"
                ],
                "summary": "Xóa văn phòng",
                "operationId": "0296608a9bfd5ece0b9744b4f864e835",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Dùng token trả về từ POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}